Clean up geocodio and google maps api logic
