+++ .github/workflows/update-data.yml
@@ -93,7 +93,7 @@ jobs:
       matrix: ${{ fromJSON(needs.matrix_prep.outputs.matrix) }}
       fail-fast: false
     env:
-      API_KEY_GOOGLE_MAPS: ${{ secrets.API_KEY_GOOGLE_MAPS }}
+      GEOCODIO_API_KEY: ${{ secrets.GEOCODIO_API_KEY }}
       GITHUB_REF: ${{ github.ref_name }} # This is changed to dev if running on a schedule
     steps:
       - name: print matrix
